requirements.txt
cryptography==3.4.8
Flask==2.2.2
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-dotenv==0.21.0
